ref: move functions to utils to be imported
