+++ packages/design-system/src/components/table/utils/index.ts
@@ -15,3 +15,4 @@
  */
 
 export { default as generateCookieTableCSV } from './generateCookieTableCSV';
+export * from './precalculatedFiltersUtils';
