+++ packages/design-system/src/components/table/utils/precalculatedFiltersUtils/index.ts
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-export { default as Table } from './components';
-export * from './persistentSettingsStore';
-export * from './useTable';
-export * from './utils';
+export { default as calculateDynamicFilterValues } from './calculateDynamicFilterValues';
+export { default as evaluateStaticFilterValues } from './evaluateStaticFilterValues';
+export { default as evaluateSelectAllOption } from './evaluateSelectAllOption';
+export { default as calculateBlockedReasonsFilterValues } from './calculateBlockedReasonsFilterValues';
